+++ packages/react/src/components/FloatingOverlay.tsx
@@ -1,4 +1,4 @@
-import {getPlatform} from '@floating-ui/utils/interactions';
+import {getPlatform} from '@floating-ui/utils/react';
 import * as React from 'react';
 import useLayoutEffect from 'use-isomorphic-layout-effect';
