+++ packages/react/src/components/FocusGuard.tsx
@@ -1,4 +1,4 @@
-import {isSafari} from '@floating-ui/utils/interactions';
+import {isSafari} from '@floating-ui/utils/react';
 import * as React from 'react';
 import useLayoutEffect from 'use-isomorphic-layout-effect';
