+++ packages/react/src/hooks/useClientPoint.ts
@@ -3,7 +3,7 @@ import {
   contains,
   getTarget,
   isMouseLikePointerType,
-} from '@floating-ui/utils/interactions';
+} from '@floating-ui/utils/react';
 import * as React from 'react';
 import useLayoutEffect from 'use-isomorphic-layout-effect';
