+++ packages/react/src/hooks/useDismiss.ts
@@ -7,7 +7,7 @@ import {
   isReactEvent,
   isVirtualClick,
   isVirtualPointerEvent,
-} from '@floating-ui/utils/interactions';
+} from '@floating-ui/utils/react';
 import * as React from 'react';
 
 import {
