+++ packages/react/src/hooks/useFocus.ts
@@ -4,7 +4,7 @@ import {
   contains,
   getDocument,
   isEventTargetWithin,
-} from '@floating-ui/utils/interactions';
+} from '@floating-ui/utils/react';
 import * as React from 'react';
 
 import type {ElementProps, FloatingContext, ReferenceType} from '../types';
