+++ packages/react/src/hooks/useTypeahead.ts
@@ -1,4 +1,4 @@
-import {stopEvent} from '@floating-ui/utils/interactions';
+import {stopEvent} from '@floating-ui/utils/react';
 import * as React from 'react';
 import useLayoutEffect from 'use-isomorphic-layout-effect';
