+++ packages/react/src/inner.ts
@@ -1,5 +1,5 @@
 import {detectOverflow, offset} from '@floating-ui/react-dom';
-import {getUserAgent} from '@floating-ui/utils/interactions';
+import {getUserAgent} from '@floating-ui/utils/react';
 import * as React from 'react';
 import {flushSync} from 'react-dom';
